+++ lib/tfmm3dlr/numthetahalf2.f
@@ -1,5 +1,5 @@
-	subroutine numthetahalf(numtets,nlams)
-	integer *4 numtets(nlams),nlams 
+       subroutine numthetahalf(numtets,nlams)
+        integer numtets(nlams),nlams 
 c
 c This routine returns the number of Fourier modes needed in the
 c phi integral for each of the discrete lambda values given
